+++ CHANGELOG.md
@@ -9,6 +9,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+### Changed
+
+### Fixed
+
+### Removed
+
+## [0.0.27] - 2021-11-12
+
+### Added
+
 - `SubEntity` decorator
 - `CountColumn` decorator
 - `EntityType` type